rename `helium` folder as `arm2d`
