+++ src/draw/sw/arm2d/lv_draw_sw_arm2d.h
